refactor extract ipcMain
