+++ index.js
@@ -1,15 +1,14 @@
 "use strict";
 const electron        = require("electron");
-const {app, BrowserWindow, Menu, ipcMain} = electron;
+const {app, BrowserWindow, Menu} = electron;
 
-const NodeStatic = require('node-static');
-const path = require('path')
-const url  = require('url')
-const http = require('http');
+const path = require('path');
+const url  = require('url');
 
 const settings = require('electron-settings');
 
-const application_menu = require('./lib/js/main_process/application-menu')
+const application_menu = require('./lib/js/main_process/application-menu');
+const ipc_main = require('./lib/js/main_process/ipc_main');
 
 global.mainWindow = null;
 
@@ -22,28 +21,7 @@ app.on('ready', () => {
     presentation: 'http://localhost:' + settings.get('server.port') + '/reveal_view.html',
     print: settings.get('url.presentation') + '?print-pdf'
   });
-  start_server(settings.get('server.port'));
-});
-
-var file = new NodeStatic.Server(__dirname + '/');
-
-function start_server(target_port) {
-  if( global.http_server != undefined && global.http_server != null) {
-    global.http_server.close( function() {
-      console.log('server is closed!!');
-    });
-  }
-  global.http_server = http.createServer(function (request, response) {
-    request.addListener('end', function () {
-      file.serve(request, response);
-    }).resume();
-  });
-  global.http_server.listen(target_port);
-}
-
-ipcMain.on('start-server', (event, arg) => {
-  start_server(settings.get('server.port'));
-  event.sender.send('start-server-reply');
+  ipc_main.start_server(settings.get('server.port'));
 });
 
 app.on('window-all-closed', () => app.quit());
