+++ lib/js/main_process/ipc_main.js
@@ -1,0 +1,30 @@
+"use strict";
+const electron  = require("electron");
+const {ipcMain} = electron;
+const settings = require('electron-settings');
+
+const NodeStatic = require('node-static');
+const http = require('http');
+
+ipcMain.on('start-server', (event, arg) => {
+  start_server(settings.get('server.port'));
+  event.sender.send('start-server-reply');
+});
+
+var file = new NodeStatic.Server(__dirname + '/');
+
+module.exports.start_server = start_server;
+function start_server(target_port) {
+  if( global.http_server != undefined && global.http_server != null) {
+    global.http_server.close( function() {
+      console.log('server is closed!!');
+    });
+  }
+  global.http_server = http.createServer(function (request, response) {
+    request.addListener('end', function () {
+      file.serve(request, response);
+    }).resume();
+  });
+  global.http_server.listen(target_port);
+}
+
